Combine query test files
