+++ tests/datastore_firestore/test_query.py
@@ -27,8 +27,9 @@ def sample_data(collection, reset_firestore):
     for x in range(1, 6):
         collection.add({"x": x})
 
+# ===== Query =====
 
-def _exercise_firestore(collection):
+def _exercise_firestore_query(collection):
     query = collection.select("x").limit(10).order_by("x").where(field_path="x", op_string="<=", value=3)
     assert len(query.get()) == 3
     assert len(list(query.stream())) == 3
@@ -54,7 +55,7 @@ def test_firestore_query(collection):
     )
     @background_task(name="test_firestore_query")
     def _test():
-        _exercise_firestore(collection)
+        _exercise_firestore_query(collection)
 
     _test()
 
@@ -68,4 +69,48 @@ def test_firestore_query_generators(collection, assert_trace_for_generator):
 @validate_database_duration()
 @background_task()
 def test_firestore_query_db_duration(collection):
-    _exercise_firestore(collection)
+    _exercise_firestore_query(collection)
+
+# ===== AggregationQuery =====
+
+def _exercise_firestore_aggregation(collection):
+    aggregation_query = collection.select("x").where(field_path="x", op_string="<=", value=3).count()
+    assert aggregation_query.get()[0][0].value == 3
+    assert list(aggregation_query.stream())[0][0].value == 3
+
+
+def test_firestore_aggregation_query(collection):
+    _test_scoped_metrics = [
+        ("Datastore/statement/Firestore/%s/stream" % collection.id, 1),
+        ("Datastore/statement/Firestore/%s/get" % collection.id, 1),
+    ]
+
+    _test_rollup_metrics = [
+        ("Datastore/operation/Firestore/get", 1),
+        ("Datastore/operation/Firestore/stream", 1),
+        ("Datastore/all", 2),
+        ("Datastore/allOther", 2),
+    ]
+    @validate_transaction_metrics(
+        "test_firestore_aggregation_query",
+        scoped_metrics=_test_scoped_metrics,
+        rollup_metrics=_test_rollup_metrics,
+        background_task=True,
+    )
+    @background_task(name="test_firestore_aggregation_query")
+    def _test():
+        _exercise_firestore_aggregation(collection)
+
+    _test()
+
+
+@background_task()
+def test_firestore_aggregation_query_generators(collection, assert_trace_for_generator):
+    aggregation_query = collection.select("x").where(field_path="x", op_string="<=", value=3).count()
+    assert_trace_for_generator(aggregation_query.stream)
+
+
+@validate_database_duration()
+@background_task()
+def test_firestore_aggregation_query_db_duration(collection):
+    _exercise_firestore_aggregation(collection)
